fix: change top heading level for public page component
